src/clickup2notion/requirements.txt
aiohttp
aiohttp-client-cache
click
notion-client
python-dotenv
